@@ -2,18 +2,21 @@ class Map < ActiveRecord::Base
 
   belongs_to :user
 
-  has_many :topicmappings, :class_name => 'Mapping', :conditions => {:category => 'Topic'}
-  has_many :synapsemappings, :class_name => 'Mapping', :conditions => {:category => 'Synapse'}
-
-  has_many :topics, :through => :topicmappings
-  has_many :synapses, :through => :synapsemappings
+  has_many :topicmappings, -> { Mapping.topicmapping }, class_name: :Mapping, dependent: :destroy
+  has_many :synapsemappings, -> { Mapping.synapsemapping }, class_name: :Mapping, dependent: :destroy
+  has_many :topics, through: :topicmappings, source: :mappable, source_type: "Topic"
+  has_many :synapses, through: :synapsemappings, source: :mappable, source_type: "Synapse"
 
   # This method associates the attribute ":image" with a file attachment
   has_attached_file :screenshot, :styles => {
    :thumb => ['188x126#', :png]
    #:full => ['940x630#', :png]
   },
-  :default_url => "/assets/missing-map.png"
+  :default_url => 'https://s3.amazonaws.com/metamaps-assets/site/missing-map.png'
+  validates :name, presence: true
+  validates :arranged, inclusion: { in: [true, false] }
+  validates :permission, presence: true
+  validates :permission, inclusion: { in: Perm::ISSIONS.map(&:to_s) }
     
   # Validate the attached image is image/jpg, image/png, etc
   validates_attachment_content_type :screenshot, :content_type => /\Aimage\/.*\Z/
@@ -23,13 +26,7 @@ class Map < ActiveRecord::Base
   end
 
   def mk_permission
-    if self.permission == "commons"
-      "co"
-    elsif self.permission == "public"
-      "pu"
-    elsif self.permission == "private"
-      "pr"
-    end
+    Perm.short(permission)
   end
 
   #return an array of the contributors to the map
@@ -44,41 +41,41 @@ class Map < ActiveRecord::Base
   end
 
   def topic_count
-    self.topics.length
+    topics.length
   end
 
   def synapse_count
-    self.synapses.length
+    synapses.length
   end
 
   def user_name
-    self.user.name
+    user.name
   end
 
   def user_image
-    self.user.image.url
+    user.image.url
   end
 
   def contributor_count 
-    self.contributors.length
+    contributors.length
   end
 
   def screenshot_url
-    self.screenshot.url(:thumb)
+    screenshot.url(:thumb)
   end
 
   def created_at_str
-    self.created_at.strftime("%m/%d/%Y")
+    created_at.strftime("%m/%d/%Y")
   end
 
   def updated_at_str
-    self.updated_at.strftime("%m/%d/%Y")
+    updated_at.strftime("%m/%d/%Y")
   end
 
   def as_json(options={})
-    json = super(:methods =>[:user_name, :user_image, :topic_count, :synapse_count, :contributor_count, :screenshot_url], :except => [:screenshot_content_type, :screenshot_file_size, :screenshot_file_name, :screenshot_updated_at, :created_at, :updated_at])
-    json[:created_at] = self.created_at_str
-    json[:updated_at] = self.updated_at_str
+    json = super(:methods =>[:user_name, :user_image, :topic_count, :synapse_count, :contributor_count, :screenshot_url], :except => [:screenshot_content_type, :screenshot_file_size, :screenshot_file_name, :screenshot_updated_at])
+    json[:created_at_clean] = created_at_str
+    json[:updated_at_clean] = updated_at_str
     json
   end
 
@@ -109,14 +106,6 @@ class Map < ActiveRecord::Base
   		return false
   	end
   	return self
-  end
-  
-  # returns Boolean if user allowed to view Topic, Synapse, or Map
-  def authorize_to_view(user)  
-  	if (self.permission == "private" && self.user != user)
-  		return false
-  	end
-  	return true
   end
 
   def decode_base64(imgBase64)
