@@ -1,10 +1,16 @@
 class Mapping < ActiveRecord::Base
 
-  belongs_to :topic, :class_name => "Topic", :foreign_key => "topic_id"
-  belongs_to :synapse, :class_name => "Synapse", :foreign_key => "synapse_id"
-  belongs_to :map, :class_name => "Map", :foreign_key => "map_id"
+  scope :topicmapping, -> { where(mappable_type: :Topic) }
+  scope :synapsemapping, -> { where(mappable_type: :Synapse) }
 
+  belongs_to :mappable, polymorphic: true
+  belongs_to :map, :class_name => "Map", :foreign_key => "map_id", touch: true
   belongs_to :user
+
+  validates :xloc, presence: true
+  validates :yloc, presence: true
+  validates :map, presence: true
+  validates :mappable, presence: true
   
   def user_name
     self.user.name
