@@ -3,13 +3,16 @@ class Topic < ActiveRecord::Base
 
   belongs_to :user
 
-  has_many :synapses1, :class_name => 'Synapse', :foreign_key => 'node1_id'
-  has_many :synapses2, :class_name => 'Synapse', :foreign_key => 'node2_id'
+  has_many :synapses1, :class_name => 'Synapse', :foreign_key => 'node1_id', dependent: :destroy
+  has_many :synapses2, :class_name => 'Synapse', :foreign_key => 'node2_id', dependent: :destroy
   has_many :topics1, :through => :synapses2, :source => :topic1
   has_many :topics2, :through => :synapses1, :source => :topic2
 
-  has_many :mappings
+  has_many :mappings, as: :mappable, dependent: :destroy
   has_many :maps, :through => :mappings
+
+  validates :permission, presence: true
+  validates :permission, inclusion: { in: Perm::ISSIONS.map(&:to_s) }
     
   # This method associates the attribute ":image" with a file attachment
   has_attached_file :image
@@ -39,27 +42,27 @@ class Topic < ActiveRecord::Base
   belongs_to :metacode
 
   def user_name
-    self.user.name
+    user.name
   end
 
   def user_image
-    self.user.image.url
+    user.image.url
   end
 
   def map_count
-    self.maps.count
+    maps.count
   end
 
   def synapse_count
-    self.synapses.count
+    synapses.count
   end
 
   def inmaps
-    self.maps.map(&:name)
+    maps.map(&:name)
   end
 
   def inmapsLinks
-    self.maps.map(&:id)
+    maps.map(&:id)
   end
 
   def as_json(options={})
@@ -71,24 +74,18 @@ class Topic < ActiveRecord::Base
   end
   
   def mk_permission
-    if self.permission == "commons"
-      "co"
-    elsif self.permission == "public"
-      "pu"
-    elsif self.permission == "private"
-      "pr"
-    end
+    Perm.short(permission)
   end
 
   # has no viewable synapses helper function
   def has_viewable_synapses(current)
   	result = false
-  	self.synapses.each do |synapse|
-  		if synapse.authorize_to_view(current)
+  	synapses.each do |synapse|
+  		if synapse.authorize_to_show(current)
   			result = true
   		end
   	end
-  	return result
+  	result
   end
   
   ##### PERMISSIONS ######
@@ -110,12 +107,11 @@ class Topic < ActiveRecord::Base
 	end
 	return self
   end
-  
-  # returns Boolean if user allowed to view Topic, Synapse, or Map
-  def authorize_to_view(user)  
-	if (self.permission == "private" && self.user != user)
-		return false
-	end
-	return true
+
+  def authorize_to_delete(user)  
+    if (self.user == user || user.admin)
+      return self
+    end
+    return false
   end
 end
