@@ -1,13 +1,13 @@
 Metamaps::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
+  config.log_level = :info
+  config.eager_load = false
+
   # In the development environment your application's code is reloaded on
   # every request. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
-
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
@@ -15,20 +15,22 @@ Metamaps::Application.configure do
   
   # S3 file storage
   config.paperclip_defaults = {
-    :storage => :s3,
-    :s3_credentials => {
-      :bucket => ENV['S3_BUCKET_NAME'],
-      :access_key_id => ENV['AWS_ACCESS_KEY_ID'],
-      :secret_access_key => ENV['AWS_SECRET_ACCESS_KEY']
-    }
+    storage: :s3,
+    s3_credentials: {
+      bucket: ENV['S3_BUCKET_NAME'],
+      access_key_id: ENV['AWS_ACCESS_KEY_ID'],
+      secret_access_key: ENV['AWS_SECRET_ACCESS_KEY']
+    },
+    s3_protocol: 'https'
   }
     
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.smtp_settings = {
-    address:              'mail.metamaps.cc',
-    port:                 587,
-    user_name:            'team@metamaps.cc',
-    password:             'RcxX+s:fht49UX',
+    address:       ENV['SMTP_SERVER'],
+    port:          ENV['SMTP_PORT'],
+    user_name:     ENV['SMTP_USERNAME'],
+    password:      ENV['SMTP_PASSWORD'],
+    #domain:        ENV['SMTP_DOMAIN']
     authentication:       'plain',
     enable_starttls_auto: true,
     openssl_verify_mode: 'none'    }
@@ -38,19 +40,6 @@ Metamaps::Application.configure do
 
   # Print deprecation notices to the Rails logger
   config.active_support.deprecation = :log
-
-  # Only use best-standards-support built into browsers
-  config.action_dispatch.best_standards_support = :builtin
-
-  # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
-
-  # Log the query plan for queries taking more than this (works
-  # with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
-
-  # Do not compress assets
-  config.assets.compress = false
 
   # Expands the lines which load the assets
   config.assets.debug = true
