@@ -1,5 +1,8 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery
+
+  before_filter :get_invite_link
+  after_action :allow_embedding
   
   # this is for global login
   include ContentHelper
@@ -9,7 +12,13 @@ class ApplicationController < ActionController::Base
   helper_method :admin?
   
   def after_sign_in_path_for(resource)
-    sign_in_url = url_for(:action => 'new', :controller => 'sessions', :only_path => false, :protocol => 'http')
+    unsafe_uri = request.env["REQUEST_URI"]
+    if unsafe_uri.starts_with?('http') && !unsafe_uri.starts_with?('https')
+      protocol = 'http'
+    else
+      protocol = 'https'
+    end
+    sign_in_url = url_for(:action => 'new', :controller => 'sessions', :only_path => false, :protocol => protocol)
 
     if request.referer == sign_in_url
       super
@@ -37,7 +46,7 @@ private
   end
     
   def require_admin
-    unless authenticated? && user.admin
+    unless authenticated? && admin?
       redirect_to root_url, notice: "You need to be an admin for that."
       return false
     end
@@ -47,13 +56,22 @@ private
     current_user
   end
   
-    
   def authenticated?
     current_user
   end
     
   def admin?
-    current_user && current_user.admin
+    authenticated? && current_user.admin
   end
-  
+
+  def get_invite_link
+    @invite_link = "#{request.base_url}/join" + (current_user ? "?code=#{current_user.code}" : "")
+  end
+
+  def allow_embedding
+    #allow all
+    response.headers.except! 'X-Frame-Options'
+    # or allow a whitelist
+    # response.headers['X-Frame-Options'] = 'ALLOW-FROM http://blog.metamaps.cc'
+  end
 end
