@@ -11,38 +11,24 @@ class MapsController < ApplicationController
     # GET /explore/featured
     # GET /explore/mapper/:id
     def index
-
-        if request.path == "/explore"
-            redirect_to activemaps_url and return
-        end
+        return redirect_to activemaps_url if request.path == "/explore"
 
         @current = current_user
-        @user = nil
         @maps = []
-        @mapperId = nil
-
-        if !params[:page] 
-            page = 1
-        else 
-            page = params[:page]
-        end
+        page = params[:page].present? ? params[:page] : 1
 
         if request.path.index("/explore/active") != nil
             @maps = Map.where("maps.permission != ?", "private").order("updated_at DESC").page(page).per(20)
             @request = "active"
-
         elsif request.path.index("/explore/featured") != nil
             @maps = Map.where("maps.featured = ? AND maps.permission != ?", true, "private").order("updated_at DESC").page(page).per(20)
             @request = "featured"
-
         elsif request.path.index('/explore/mine') != nil  # looking for maps by me
-            if !authenticated?
-                redirect_to activemaps_url and return
-            end
+            return redirect_to activemaps_url if !authenticated?
+
             # don't need to exclude private maps because they all belong to you
             @maps = Map.where("maps.user_id = ?", @current.id).order("updated_at DESC").page(page).per(20)
             @request = "you"
-
         elsif request.path.index('/explore/mapper/') != nil  # looking for maps by a mapper
             @user = User.find(params[:id])
             @maps = Map.where("maps.user_id = ? AND maps.permission != ?", @user.id, "private").order("updated_at DESC").page(page).per(20)
@@ -73,14 +59,10 @@ class MapsController < ApplicationController
         respond_to do |format|
             format.html { 
                 @allmappers = @map.contributors
-                @alltopics = @map.topics.delete_if {|t| t.permission == "private" && (!authenticated? || (authenticated? && @current.id != t.user_id)) }
-                @allsynapses = @map.synapses.delete_if {|s| s.permission == "private" && (!authenticated? || (authenticated? && @current.id != s.user_id)) }
-                @allmappings = @map.mappings.delete_if {|m| 
-                    if m.category == "Synapse"
-                        object = m.synapse
-                    elsif m.category == "Topic"
-                        object = m.topic
-                    end
+                @alltopics = @map.topics.to_a.delete_if {|t| t.permission == "private" && (!authenticated? || (authenticated? && @current.id != t.user_id)) }
+                @allsynapses = @map.synapses.to_a.delete_if {|s| s.permission == "private" && (!authenticated? || (authenticated? && @current.id != s.user_id)) }
+                @allmappings = @map.mappings.to_a.delete_if {|m| 
+                    object = m.mappable
                     !object || (object.permission == "private" && (!authenticated? || (authenticated? && @current.id != object.user_id)))
                 }
 
@@ -101,14 +83,10 @@ class MapsController < ApplicationController
         end
 
         @allmappers = @map.contributors
-        @alltopics = @map.topics.delete_if {|t| t.permission == "private" && (!authenticated? || (authenticated? && @current.id != t.user_id)) }
-        @allsynapses = @map.synapses.delete_if {|s| s.permission == "private" && (!authenticated? || (authenticated? && @current.id != s.user_id)) }
-        @allmappings = @map.mappings.delete_if {|m| 
-            if m.category == "Synapse"
-                object = m.synapse
-            elsif m.category == "Topic"
-                object = m.topic
-            end
+        @alltopics = @map.topics.to_a.delete_if {|t| t.permission == "private" && (!authenticated? || (authenticated? && @current.id != t.user_id)) }
+        @allsynapses = @map.synapses.to_a.delete_if {|s| s.permission == "private" && (!authenticated? || (authenticated? && @current.id != s.user_id)) }
+        @allmappings = @map.mappings.to_a.delete_if {|m| 
+            object = m.mappable
             !object || (object.permission == "private" && (!authenticated? || (authenticated? && @current.id != object.user_id)))
         }
 
@@ -141,7 +119,6 @@ class MapsController < ApplicationController
 
     # POST maps
     def create
-
         @user = current_user
         @map = Map.new()
         @map.name = params[:name]
@@ -149,42 +126,45 @@ class MapsController < ApplicationController
         @map.permission = params[:permission]
         @map.user = @user
         @map.arranged = false 
-        @map.save     
 
         if params[:topicsToMap]
             @all = params[:topicsToMap]
             @all = @all.split(',')
             @all.each do |topic|
                 topic = topic.split('/')
-                @mapping = Mapping.new()
-                @mapping.category = "Topic"
-                @mapping.user = @user
-                @mapping.map  = @map
-                @mapping.topic = Topic.find(topic[0])
-                @mapping.xloc = topic[1]
-                @mapping.yloc = topic[2]
-                @mapping.save
+                mapping = Mapping.new()
+                mapping.user = @user
+                mapping.mappable = Topic.find(topic[0])
+                mapping.xloc = topic[1]
+                mapping.yloc = topic[2]
+                @map.topicmappings << mapping
+                mapping.save
             end
 
             if params[:synapsesToMap]
                 @synAll = params[:synapsesToMap]
                 @synAll = @synAll.split(',')
                 @synAll.each do |synapse_id|
-                    @mapping = Mapping.new()
-                    @mapping.category = "Synapse"
-                    @mapping.user = @user
-                    @mapping.map = @map
-                    @mapping.synapse = Synapse.find(synapse_id)
-                    @mapping.save
+                    mapping = Mapping.new()
+                    mapping.user = @user
+                    mapping.map = @map
+                    mapping.mappable = Synapse.find(synapse_id)
+                    @map.synapsemappings << mapping
+                    mapping.save
                 end
             end
 
             @map.arranged = true
-            @map.save      
         end
 
-        respond_to do |format|
+        if @map.save
+          respond_to do |format|
             format.json { render :json => @map }
+          end
+        else
+          respond_to do |format|
+            format.json { render :json => "invalid params" }
+          end
         end
     end
 
@@ -196,7 +176,7 @@ class MapsController < ApplicationController
         respond_to do |format|
             if !@map 
                 format.json { render json: "unauthorized" }
-            elsif @map.update_attributes(params[:map])
+            elsif @map.update_attributes(map_params)
                 format.json { head :no_content }
             else
                 format.json { render json: @map.errors, status: :unprocessable_entity }
@@ -234,15 +214,7 @@ class MapsController < ApplicationController
 
         @map = Map.find(params[:id]).authorize_to_delete(@current)
 
-        if @map 
-            @mappings = @map.mappings
-
-            @mappings.each do |mapping| 
-                mapping.delete
-            end
-
-            @map.delete
-        end
+        @map.delete if @map
 
         respond_to do |format|
             format.json { 
@@ -253,5 +225,12 @@ class MapsController < ApplicationController
                 end
             }
         end
+    end
+
+    private
+
+    # Never trust parameters from the scary internet, only allow the white list through.
+    def map_params
+      params.require(:map).permit(:id, :name, :arranged, :desc, :permission, :user_id)
     end
 end
