@@ -1,9 +1,8 @@
 class UsersController < ApplicationController
-
   before_filter :require_user, only: [:edit, :update, :updatemetacodes]
     
   respond_to :html, :json 
-  
+
   # GET /users/1.json
   def show
     @user = User.find(params[:id])
@@ -14,7 +13,6 @@ class UsersController < ApplicationController
   # GET /users/:id/edit
   def edit
     @user = current_user
-    
     respond_with(@user)  
   end
   
@@ -22,9 +20,9 @@ class UsersController < ApplicationController
   def update
     @user = current_user
 
-    if params[:user][:password] == "" && params[:user][:password_confirmation] == ""
+    if user_params[:password] == "" && user_params[:password_confirmation] == ""
       # not trying to change the password
-      if @user.update_attributes(params[:user])
+      if @user.update_attributes(user_params.except(:password, :password_confirmation))
         if params[:remove_image] == "1"
           @user.image = nil
         end
@@ -43,7 +41,7 @@ class UsersController < ApplicationController
       # trying to change the password
       correct_pass = @user.valid_password?(params[:current_password])
 
-      if correct_pass && @user.update_attributes(params[:user])
+      if correct_pass && @user.update_attributes(user_params)
         if params[:remove_image] == "1"
           @user.image = nil
         end
@@ -98,4 +96,9 @@ class UsersController < ApplicationController
     end
   end
 
+  private
+
+  def user_params
+    params.require(:user).permit(:name, :email, :image, :password, :password_confirmation)
+  end
 end
