@@ -1,24 +1,36 @@
 class Synapse < ActiveRecord::Base
-
   belongs_to :user
 
   belongs_to :topic1, :class_name => "Topic", :foreign_key => "node1_id"
   belongs_to :topic2, :class_name => "Topic", :foreign_key => "node2_id"
 
-  has_many :mappings
+  has_many :mappings, as: :mappable, dependent: :destroy
   has_many :maps, :through => :mappings
 
+  validates :desc, length: { minimum: 0, allow_nil: false }
+
+  validates :permission, presence: true
+  validates :permission, inclusion: { in: Perm::ISSIONS.map(&:to_s) }
+
+  validates :category, inclusion: { in: ['from-to', 'both'], allow_nil: true }
+
+  # :nocov:
   def user_name
-    self.user.name
+    user.name
   end
+  # :nocov:
 
+  # :nocov:
   def user_image
-    self.user.image.url
+    user.image.url
   end
+  # :nocov:
 
+  # :nocov:
   def as_json(options={})
     super(:methods =>[:user_name, :user_image])
   end
+  # :nocov:
   
   ##### PERMISSIONS ######
   
@@ -39,13 +51,11 @@ class Synapse < ActiveRecord::Base
 	end
 	return self
   end
-  
-  # returns Boolean if user allowed to view Topic, Synapse, or Map
-  def authorize_to_view(user)  
-	if (self.permission == "private" && self.user != user)
-		return false
-	end
-	return true
-  end
 
+  def authorize_to_delete(user)  
+    if (self.user == user || user.admin)
+      return self
+    end
+    return false
+  end
 end
