@@ -7,12 +7,10 @@ class User < ActiveRecord::Base
   has_many :maps
   has_many :mappings
 
-  before_create :generate_code
+  after_create :generate_code
 
   devise :database_authenticatable, :recoverable, :rememberable, :trackable, :registerable
   
-  attr_accessible :name, :email, :image, :password, :password_confirmation, :code, :joinedwithcode, :remember_me
-
   serialize :settings, UserPreference
 	
   validates :password, :presence => true,
@@ -37,38 +35,45 @@ class User < ActiveRecord::Base
    :ninetysix => ['96x96#', :png],
    :onetwentyeight => ['128x128#', :png]
   },
-  :default_url => "/assets/user.png"
+  :default_url => 'https://s3.amazonaws.com/metamaps-assets/site/user.png'
     
   # Validate the attached image is image/jpg, image/png, etc
   validates_attachment_content_type :image, :content_type => /\Aimage\/.*\Z/
 
+  # override default as_json
   def as_json(options={})
     { :id => self.id,
       :name => self.name,
-      :image => self.image.url(:sixtyfour)
+      :image => self.image.url(:sixtyfour),
+      :admin => self.admin
     }
   end
+
+  def as_json_for_autocomplete
+    json = {}
+    json['id'] = id
+    json['label'] = name
+    json['value'] = name
+    json['profile'] = image.url(:sixtyfour)
+    json['mapCount'] = maps.count
+    json['generation'] = generation
+    json['created_at'] = created_at.strftime("%m/%d/%Y")
+    json['rtype'] = "mapper"
+    json
+  end
   
+  #generate a random 8 letter/digit code that they can use to invite people
   def generate_code
-    #generate a random 8 letter/digit code that they can use to invite people
-	  self.code = rand(36**8).to_s(36)
-
+	  self.code ||= rand(36**8).to_s(36)
     $codes.push(self.code)
-
-    self.generation = self.get_generation
+    self.generation = get_generation!
   end
 
-  def get_generation
-    if self.joinedwithcode == self.code
-      # if your joinedwithcode equals your code you must be GEN 0
-      gen = 0
-    elsif self.generation
-      # if your generation has already been calculated then just return that value
-      gen = self.generation
+  def get_generation!
+    if code == joinedwithcode
+      update(generation: 0)
     else
-      # if your generation hasn't been calculated, base it off the
-      # generation of the person whose code you joined with + 1
-      gen = User.find_by_code(self.joinedwithcode).get_generation + 1
+      update(generation: User.find_by_code(joinedwithcode).generation + 1)
     end
   end
   
@@ -76,13 +81,12 @@ class User < ActiveRecord::Base
     # make sure we always return a UserPreference instance
     if read_attribute(:settings).nil?
       write_attribute :settings, UserPreference.new
-      read_attribute :settings
-    else
-      read_attribute :settings
     end
+    read_attribute :settings
   end
   
   def settings=(val)
     write_attribute :settings, val
   end
+
 end
