@@ -1,13 +1,12 @@
 require File.expand_path('../boot', __FILE__)
 
 require 'rails/all'
+require 'dotenv'
 
-if defined?(Bundler)
-  # If you precompile assets before deploying to production, use this line
-  Bundler.require(*Rails.groups(:assets => %w(development test)))
-  # If you want your assets lazily compiled in production, use this line
-  # Bundler.require(:default, :assets, Rails.env)
-end
+Bundler.require(*Rails.groups)
+
+ENV["RAILS_ENV"] ||= ENV["RACK_ENV"] || "development"
+Dotenv.load ".env.#{ENV["RAILS_ENV"]}", '.env'
 
 module Metamaps
   class Application < Rails::Application
@@ -16,7 +15,7 @@ module Metamaps
     # -- all .rb files in that directory are automatically loaded.
 
     # Custom directories with classes and modules you want to be autoloadable.
-    # config.autoload_paths += %W(#{config.root}/extras)
+    config.autoload_paths << Rails.root.join('app', 'services')
 
     # Only load the plugins named here, in the order given (default is alphabetical).
     # :all can be used as a placeholder for all plugins not explicitly named.
@@ -44,17 +43,15 @@ module Metamaps
     # like if you have constraints or database-specific column types
     # config.active_record.schema_format = :sql
 
-    # Enforce whitelist mode for mass assignment.
-    # This will create an empty whitelist of attributes available for mass-assignment for all models
-    # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
-    # parameters by using an attr_accessible or attr_protected declaration.
-    # config.active_record.whitelist_attributes = true
-
     # Enable the asset pipeline
-    config.assets.enabled = true
     config.assets.initialize_on_precompile = false
       
     # Version of your assets, change this if you want to expire all your assets
     config.assets.version = '2.0'
+
+    config.generators do |g|
+      g.test_framework :rspec
+    end
+    config.active_record.raise_in_transactional_callbacks = true
   end
 end
